Fixed error

+++ micro_ros_raspberrypi_pico_sdk/pico_controller.cpp
@@ -78,7 +78,7 @@ void subscription_callback_servo(const void * msgin) {
 
 void subscription_callback_motor(const void *msgin) {
     const std_msgs__msg__String *msg = (const std_msgs__msg__String *)msgin;
-    double speed = std::stoi(msg->data.data);
+    double speed = std::stod(msg->data.data);
     // Set the motor speed
     motor.getSpeed();
     motor.pidController(speed);
